requirements.txt
# Orchestration & Database Providers
apache-airflow-providers-postgres
apache-airflow-providers-amazon

# Extraction Tools
pandas
requests
boto3

# Transformation (dbt)
dbt-postgres